[feat] 지난주 당첨 번호와 몇개 일치했는지 통계 생성하는 클래스 테스트 추가 및 일치한 결과를 저장하는 랭크 클래스 생성

+++ src/main/java/step2/HitCount.java
@@ -2,16 +2,16 @@ package step2;
 
 import java.util.Objects;
 
-public class HitCount {
+public class HitCount{
 
-    private long hitCount;
+    private int hitCount;
     private boolean bonusBallHit;
 
-    public HitCount(long hitCount) {
+    public HitCount(int hitCount) {
         this(hitCount,false);
     }
 
-    public HitCount(long hitCount,boolean bonusBallHit) {
+    public HitCount(int hitCount,boolean bonusBallHit) {
         this.hitCount = hitCount;
         this.bonusBallHit = bonusBallHit;
     }
